requirementsnew.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
anndata==0.10.7
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.6
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autopage==0.5.2
Babel==2.14.0
beautifulsoup4==4.12.3
bgflow @ git+https://github.com/noegroup/bgflow.git@fbba56fac3eb88f6825d2bd4f745ee75ae9715e1
biopython==1.83
bleach==6.1.0
boltzgen @ git+https://github.com/VincentStimper/boltzmann-generators.git@2b177fc155f533933489b8fce8d6483ebad250d3
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
certifi==2024.2.2
cffi==1.16.0
cfgv @ file:///home/conda/feedstock_root/build_artifacts/cfgv_1629909281805/work
charset-normalizer==3.3.2
chunkgfn @ file:///home/mila/l/lena-nehale.ezzine/Chunk-GFN
clean-fid==0.1.35
click==8.1.7
cliff==4.6.0
cloudpickle==3.0.0
cmaes==0.10.0
cmd2==2.4.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
# Editable install with no version control (dem==0.0.1)
-e /home/mila/l/lena-nehale.ezzine/Amgen/cremp/DEM
distlib @ file:///home/conda/feedstock_root/build_artifacts/distlib_1702383208639/work
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fab @ git+https://github.com/jarridrb/fab-torch.git@febeeaa521eb5793f65b246a82e9dbd1c745c00e
fastjsonschema==2.19.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1714422806336/work
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1713794390601/work
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
hydra-colorlog==1.2.0
hydra-core==1.3.0
hydra-joblib-launcher==1.2.0
hydra-optuna-sweeper==1.2.0
hydra-submitit-launcher==1.2.0
identify @ file:///home/conda/feedstock_root/build_artifacts/identify_1713672996230/work
idna==3.6
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1711040877059/work
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1673103042956/work
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
larsflow @ git+https://github.com/VincentStimper/resampled-base-flows.git@18db5bf28ffa1d5ab9ef2b63856e186affee604b
legacy-api-wrap==1.4
lightning==2.2.1
lightning-bolts==0.7.0
lightning-utilities==0.10.1
llvmlite==0.42.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.2
nest-asyncio==1.6.0
networkx==3.2.1
nflows==0.14
nodeenv @ file:///home/conda/feedstock_root/build_artifacts/nodeenv_1683892983968/work
normflows==1.7.3
notebook==7.1.1
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
OpenMM==8.1.1
optuna==2.10.1
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pastel==0.2.1
pathtools==0.1.2
patsy==0.5.6
pbr==6.0.0
periodictable==1.5.2
pexpect==4.9.0
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1713912794367/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1713667077545/work
poethepoet==0.10.0
polyleven==0.8
POT==0.9.3
pre_commit @ file:///home/conda/feedstock_root/build_artifacts/pre-commit_1711480319650/work
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pynndescent==0.5.12
pyparsing==3.1.2
pyperclip==1.8.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest @ file:///home/conda/feedstock_root/build_artifacts/pytest_1704035161844/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
residual_flows @ git+https://github.com/VincentStimper/residual-flows.git@2bbbf70570f1ce4ec8de21da3423fcc773f48f98
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rootutils==1.0.7
rpds-py==0.18.0
scanpy==1.10.1
scikit-learn==1.4.2
scipy==1.12.0
scprep==1.2.3
seaborn==0.13.2
Send2Trash==1.8.2
sentry-sdk==1.42.0
session_info==1.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
statsmodels==0.14.2
stdlib-list==0.10.0
stevedore==5.2.0
submitit==1.5.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.5
torch==2.2.1
torchcde==0.2.5
torchcfm==1.0.5
torchdiffeq==0.2.3
torchdyn==1.0.6
torchgfn==1.1.1
torchmetrics==1.3.1
torchsde==0.2.6
torchtyping==0.1.4
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trampoline==0.1.2
triton==2.2.0
typeguard==4.1.5
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
tzdata==2024.1
ukkonen @ file:///home/conda/feedstock_root/build_artifacts/ukkonen_1695549355893/work
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.2.1
virtualenv @ file:///home/conda/feedstock_root/build_artifacts/virtualenv_1714439372989/work
wandb==0.16.4
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wurlitzer @ file:///home/conda/feedstock_root/build_artifacts/wurlitzer_1718203955352/work
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
